pyproject.toml
[project]
name = "rag-evaluator"
version = "0.1.0"
description = "A GQM-based Framework for RAG System Evaluation"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "langchain>=0.3.25",
    "numpy>=2.0.2",
    "pinecone>=7.3.0",
    "google-generativeai>=0.8.5",
    "bert-score>=0.3.13",
    "pydantic>=2.11.4",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.7",
    "voyageai>=0.3.2",
    "python-dotenv>=1.1.0",
    "transformers>=4.49.0",
    "spacy-transformers>=1.3.9",
    "accelerate>=0.26.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "python-multipart>=0.0.9",
]